[#173171643] refactoring
