+++ ts/components/messages/MessageListCTABar.tsx
@@ -2,7 +2,7 @@ import { fromNullable, Option } from "fp-ts/lib/Option";
 import { capitalize } from "lodash";
 import { Text, View } from "native-base";
 import React from "react";
-import { Linking, StyleSheet } from "react-native";
+import { StyleSheet } from "react-native";
 import { connect } from "react-redux";
 import { CreatedMessageWithContent } from "../../../definitions/backend/CreatedMessageWithContent";
 import { ServicePublic } from "../../../definitions/backend/ServicePublic";
@@ -13,17 +13,13 @@ import { CTA } from "../../types/MessageCTA";
 import { formatDateAsDay, formatDateAsMonth } from "../../utils/dates";
 import {
   getCTA,
+  handleCtaAction,
   hasCTAValidActions,
   isCtaActionValid,
   isExpired,
   paymentExpirationInfo
 } from "../../utils/messages";
 import ButtonDefaultOpacity from "../ButtonDefaultOpacity";
-import {
-  getInternalRoute,
-  handleInternalLink
-} from "../ui/Markdown/handlers/internalLink";
-import { deriveCustomHandledLink } from "../ui/Markdown/handlers/link";
 import CalendarEventButton from "./CalendarEventButton";
 import CalendarIconComponent from "./CalendarIconComponent";
 import PaymentButton from "./PaymentButton";
@@ -135,18 +131,6 @@ class MessageListCTABar extends React.PureComponent<Props> {
     });
   }
 
-  private handleCTAAction = (cta: CTA) => {
-    const maybeInternalLink = getInternalRoute(cta.action);
-    if (maybeInternalLink.isSome()) {
-      handleInternalLink(this.props.dispatch, cta.action);
-    } else {
-      const maybeHandledAction = deriveCustomHandledLink(cta.action);
-      if (maybeHandledAction.isSome()) {
-        Linking.openURL(maybeHandledAction.value).catch(() => 0);
-      }
-    }
-  };
-
   private renderCTA(cta?: CTA, primary: boolean = false) {
     if (cta === undefined || !isCtaActionValid(cta)) {
       return null;
@@ -157,7 +141,7 @@ class MessageListCTABar extends React.PureComponent<Props> {
         primary={primary}
         disabled={false}
         bordered={!primary}
-        onPress={() => this.handleCTAAction(cta)}
+        onPress={() => handleCtaAction(cta, this.props.dispatch)}
         xsmall={true}
         style={{
           flex: 1
