+++ ts/utils/messages.ts
@@ -10,12 +10,17 @@ import {
   some
 } from "fp-ts/lib/Option";
 import FM from "front-matter";
+import { Linking } from "react-native";
+import { Dispatch } from "redux";
 import { CreatedMessageWithContent } from "../../definitions/backend/CreatedMessageWithContent";
 import { CreatedMessageWithContentAndAttachments } from "../../definitions/backend/CreatedMessageWithContentAndAttachments";
 import { MessageBodyMarkdown } from "../../definitions/backend/MessageBodyMarkdown";
 import { PrescriptionData } from "../../definitions/backend/PrescriptionData";
 import { Locales } from "../../locales/locales";
-import { getInternalRoute } from "../components/ui/Markdown/handlers/internalLink";
+import {
+  getInternalRoute,
+  handleInternalLink
+} from "../components/ui/Markdown/handlers/internalLink";
 import { deriveCustomHandledLink } from "../components/ui/Markdown/handlers/link";
 import I18n, { translations } from "../i18n";
 import { CTA, CTAS, MessageCTA } from "../types/MessageCTA";
@@ -53,6 +58,18 @@ export function messageNeedsCTABar(
     getCTA(message).isSome()
   );
 }
+
+export const handleCtaAction = (cta: CTA, dispatch: Dispatch) => {
+  const maybeInternalLink = getInternalRoute(cta.action);
+  if (maybeInternalLink.isSome()) {
+    handleInternalLink(dispatch, cta.action);
+  } else {
+    const maybeHandledAction = deriveCustomHandledLink(cta.action);
+    if (maybeHandledAction.isSome()) {
+      Linking.openURL(maybeHandledAction.value).catch(() => 0);
+    }
+  }
+};
 
 export const hasPrescriptionData = (
   message: CreatedMessageWithContentAndAttachments
